+++ daemon/src/ctrl_gfx/gfx.rs
@@ -2,10 +2,7 @@ use ctrl_gfx::error::GfxError;
 use ctrl_gfx::*;
 use ctrl_rog_bios::CtrlRogBios;
 use log::{error, info, warn};
-use logind_zbus::{
-    types::{SessionClass, SessionInfo, SessionType},
-    ManagerProxy, SessionProxy,
-};
+use logind_zbus::{ManagerProxy, SessionProxy, types::{SessionClass, SessionInfo, SessionState, SessionType}};
 use rog_types::gfx_vendors::{GfxRequiredUserAction, GfxVendors};
 use std::sync::mpsc;
 use std::{io::Write, ops::Add, path::Path, time::Instant};
@@ -326,6 +323,12 @@ impl CtrlGraphics {
                 {
                     return Ok(());
                 }
+                if output
+                    .stderr
+                    .ends_with("is builtin.\n".as_bytes())
+                {
+                    return Err(GfxError::VfioBuiltin.into());
+                }
                 if output.stderr.ends_with("Permission denied\n".as_bytes()) {
                     warn!(
                         "{} {} failed: {:?}",
@@ -335,6 +338,9 @@ impl CtrlGraphics {
                     );
                     warn!("GFX: It may be safe to ignore the above error, run `lsmod |grep {}` to confirm modules loaded", driver);
                     return Ok(());
+                }
+                if String::from_utf8_lossy(&output.stderr).contains(&format!("Module {} not found", driver)) {
+                    return Err(GfxError::MissingModule(driver.into()).into());
                 }
                 if count >= MAX_TRIES {
                     let msg = format!(
@@ -350,7 +356,7 @@ impl CtrlGraphics {
             }
 
             count += 1;
-            std::thread::sleep(std::time::Duration::from_millis(250));
+            std::thread::sleep(std::time::Duration::from_millis(50));
         }
     }
 
@@ -455,7 +461,6 @@ impl CtrlGraphics {
                 Self::unbind_remove_nvidia(&devices)?;
             }
         }
-
         Ok(())
     }
 
@@ -468,9 +473,10 @@ impl CtrlGraphics {
             if session_proxy.get_class()? == SessionClass::User {
                 match session_proxy.get_type()? {
                     SessionType::X11 | SessionType::Wayland | SessionType::MIR => {
-                        //if session_proxy.get_active()? {
-                        return Ok(true);
-                        //}
+                        match session_proxy.get_state()? {
+                            SessionState::Online | SessionState::Active => return Ok(true),
+                            SessionState::Closing | SessionState::Invalid => {}
+                        }
                     }
                     _ => {}
                 }
@@ -499,8 +505,7 @@ impl CtrlGraphics {
         loop {
             let tmp = manager.list_sessions()?;
             if !tmp.iter().eq(&sessions) {
-                warn!("GFX: Sessions list changed");
-                warn!("GFX: Old list:\n{:?}\nNew list:\n{:?}", &sessions, &tmp);
+                info!("GFX thread: Sessions list changed");
                 sessions = tmp;
             }
 
@@ -523,30 +528,15 @@ impl CtrlGraphics {
             sleep(SLEEP_PERIOD);
         }
 
-        info!("GFX: all graphical user sessions ended, continuing");
+        info!("GFX thread: all graphical user sessions ended, continuing");
         Self::do_display_manager_action("stop")?;
-
-        match Self::wait_display_manager_state("inactive") {
-            Ok(_) => info!("GFX: display-manager stopped"),
-            Err(err) => {
-                warn!("GFX: {}", err);
-                warn!("GFX: Retry stop display manager");
-                Self::do_display_manager_action("stop")?;
-                Self::wait_display_manager_state("inactive")?;
-            }
-        }
+        Self::wait_display_manager_state("inactive")?;
 
         Self::do_vendor_tasks(vendor, &devices, &bus)?;
-        Self::do_display_manager_action("start")?;
-
-        if Self::wait_display_manager_state("active").is_err() {
-            error!("GFX: display-manager failed to start normally, attempting restart");
-            Self::do_display_manager_action("restart")?;
-            Self::wait_display_manager_state("active")?;
-        }
+        Self::do_display_manager_action("restart")?;
         // Save selected mode in case of reboot
         Self::save_gfx_mode(vendor, config);
-        info!("GFX: display-manager started");
+        info!("GFX thread: display-manager started");
 
         let v: &str = vendor.into();
         info!("GFX: Graphics mode changed to {} successfully", v);
@@ -560,7 +550,7 @@ impl CtrlGraphics {
                 info!("GFX: Cancelling previous thread");
                 tx.send(true)
                     .map_err(|err| {
-                        warn!("GFX: {}", err);
+                        warn!("GFX thread: {}", err);
                     })
                     .ok();
             }